Highlight searchterms in mkdocs projects
Sphinx has a theme agnostic feature to highlight terms given by the
"highlight" GET parameter. This is used by RTD when linking from the search
page to the documentation page.

This did not work for mkdocs projects yet. We add the logic for this to
readthedocs-embed-js in order to make this feature available independently of
the used mkdocs theme.